fix(sale_product_multi_add): Correct email error in translation file and optimize price calculation logic
Fix incorrect email address in zh_CN.po file
Remove redundant price calculation code in sale_import_products.py, switch to system default price calculation logic
Add test cases to verify price list discount functionality
修复翻译文件中的电子邮件地址错误，删除不必要的测试文件，并更新导入产品向导的行值计算逻辑
