+++ sale_product_multi_add/i18n/zh_CN.po
@@ -7,7 +7,7 @@ msgstr ""
 "Project-Id-Version: Odoo Server 14.0\n"
 "Report-Msgid-Bugs-To: \n"
 "PO-Revision-Date: 2021-03-11 10:45+0000\n"
-"Last-Translator: 黎伟杰 <674416404@qq.com>，何修宝 <286522352@qq.ccom>\n"
+"Last-Translator: 黎伟杰 <674416404@qq.com>，何修宝 <286522352@qq.com>\n"
 "Language-Team: none\n"
 "Language: zh_CN\n"
 "MIME-Version: 1.0\n"
